float tensor

+++ CNN/eval_francia.py
@@ -64,7 +64,7 @@ def main():
         with torch.no_grad():
             for data in traces:
 
-                data = torch.from_numpy(data)
+                data = torch.from_numpy(data).float()
                 data = data.to(device)
                 outputs = net(data)
